fix: Correct Clerk client usage and database type mismatch
- Fix Clerk client import pattern to match working examples
- Update search function return types from REAL to DOUBLE PRECISION
- Resolves TypeError in user space lookup and database type errors
- Both search and resources APIs now properly handle user spaces

+++ database/search-function.sql
@@ -17,8 +17,8 @@ RETURNS TABLE (
   created_by UUID,
   updated_at TIMESTAMPTZ,
   created_at TIMESTAMPTZ,
-  rank REAL,
-  score REAL
+  rank DOUBLE PRECISION,
+  score DOUBLE PRECISION
 ) AS $$
 BEGIN
   RETURN QUERY
